requirements.txt
absl-py==1.2.0
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
azure-common==1.1.28
azure-core==1.25.0
azure-identity==1.10.0
azure-mgmt-core==1.3.1
azure-mgmt-media==10.0.0
backcall==0.2.0
beautifulsoup4==4.11.1
bestcaptchasolverapi3==0.2
boto3==1.24.47
botocore==1.27.47
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cramjam==2.5.0
cryptography==37.0.4
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastparquet==0.8.1
flatbuffers==1.12
fonttools==4.34.4
fpdf==1.7.2
fsspec==2022.7.1
gast==0.4.0
google-api-core==2.10.0
google-api-python-client==2.60.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
google-pasta==0.2.0
googleapis-common-protos==1.56.4
grpcio==1.48.1
h11==0.13.0
h5py==3.7.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
isodate==0.6.1
jedi==0.18.1
jmespath==1.0.1
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.11.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mlxtend==0.20.0
msal==1.18.0
msal-extensions==1.0.0
msedge-selenium-tools==3.141.4
msrest==0.7.1
nest-asyncio==1.5.5
numpy==1.23.1
oauthlib==3.2.0
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
portalocker==2.5.1
prompt-toolkit==3.0.30
protobuf==4.21.5
psutil==5.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyodbc==4.0.34
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-power-bi==0.1.2
pytz==2022.1
pyzmq==23.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.1.2
scipy==1.9.0
seaborn==0.11.2
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.3.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.2
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
Unidecode==1.3.4
unzip==1.0.0
uritemplate==4.1.1
urllib3==1.26.10
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.1.0
zipp==3.8.1